Extract wave request router models
